Split fraud cleanup helpers
