+++ frontend/src/server/fraud-cleanup/constants.ts
@@ -1,0 +1,7 @@
+export const RESTRICTED_ACCOUNT_MESSAGE =
+  'Your account is temporarily restricted for security reasons. Please contact support@maxvideoai.com.';
+
+export const CREDIT_REVERSAL_DESCRIPTION = 'Credit reversal: refunded suspected fraudulent payment';
+export const FRAUD_REVERSAL_REASON = 'fraud_credit_reversal';
+export const DEFAULT_REPEATED_SMALL_TOPUP_THRESHOLD = 3;
+export const DEFAULT_SMALL_TOPUP_MAX_CENTS = 1500;
